Use `size_t`, not `int`, to track sizes.

+++ lib/xmalloc.c
@@ -29,7 +29,7 @@
 
 typedef struct hashTableItemRec {
   void *ptr;
-  int bytes;
+  size_t bytes;
   const char *file;
   int line;
   const char *func;
@@ -90,7 +90,7 @@ hash_void_ptr(void *ptr)
 }
 
 static void
-hash_table_add(hashTable *tbl, void *ptr, int bytes,
+hash_table_add(hashTable *tbl, void *ptr, size_t bytes,
 	       const char *file, int line, const char *func)
 {
   int i;
@@ -199,9 +199,9 @@ xmalloc_configure(int fail_after)
 int
 xmalloc_dump_leaks(void)
 {
-  int i;
-  int num_leaks = 0;
-  int leaked_bytes = 0;
+  unsigned int i;
+  unsigned int num_leaks = 0;
+  size_t leaked_bytes = 0;
   hashTableItem *item;
 
   xmalloc_init();
@@ -211,7 +211,7 @@ xmalloc_dump_leaks(void)
       item = xmalloc_table->table[i];
       while (item != NULL)
 	{
-	  printf("%s:%d: %s: %d bytes at %p not freed\n",
+	  printf("%s:%d: %s: %zu bytes at %p not freed\n",
 		 item->file, item->line, item->func, item->bytes, item->ptr);
 	  num_leaks++;
 	  leaked_bytes += item->bytes;
@@ -221,7 +221,7 @@ xmalloc_dump_leaks(void)
   if (num_leaks == 0)
     printf("No memory leaks.\n");
   else
-    printf("%d unfreed memory chuncks, total %d unfreed bytes.\n",
+    printf("%u unfreed memory chuncks, total %zu unfreed bytes.\n",
 	   num_leaks, leaked_bytes);
   printf("Peak memory consumption %d bytes (%.1f kB, %.1f MB) in %d blocks ",
 	 xmalloc_peak, (double)xmalloc_peak / 1024,
